+++ src/provider/parser/sqlBlcok.ts
@@ -1,0 +1,6 @@
+import { Range } from "vscode";
+
+export class SQLBlock {
+    sql: string;
+    range: Range;
+}
